% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive.R
\name{chessPI.archives}
\alias{chessPI.archives}
\title{Get Archive List}
\usage{
chessPI.archives(username = "Finished_Milk")
}
\arguments{
\item{username}{The Chess.com username.}
}
\value{
An array of available archive dates.
}
\description{
Chess.com organizes its archive of your games by date.
This function returns the dates that can be used in the range field of the pull
function to narrow down your records.
Leaving the input blank will pull my games.
}
\examples{
chessPI.archives("Finished_Milk")
chessPI.archives()

}
